% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{A simulated dataset for Session 7 of the MLW / COM Statistics and R workshop.}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{type}{Indicates the type of observation, either A or B.}
\item{x}{A measurement x.}
\item{y}{Another measurement y.}
\item{z}{Yet another measurement z.}
}
}
\source{
Generated by Marc Henrion as part of the Session 7 exercise.
}
\usage{
dat
}
\description{
A simulated dataset for Session 7 of the MLW / COM Statistics and R workshop.
}
\keyword{datasets}
