% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doAnalysis}
\alias{doAnalysis}
\title{Does the analysis for the exercise for Session 6 from the MLW / KUHeS Statistics and R workshop.}
\usage{
doAnalysis(analysis)
}
\arguments{
\item{analysis}{A character string indicating which analysis to do. Needs to be one of "t.test" (for a t-test), "linReg" (for a linear regression).}
}
\value{
An object of class \code{htest} or \code{glm} depending whether a t-test or a linear regression was performed
}
\description{
This is an example package for Session 6 of the MLW / KUHeS Statistics and R workshop. This function performs either a t-test or a linear regression on a simulated, internally stored dataset, dat.
}
\examples{
## Example 1 - t-test
doAnalysis("t.test")

## Example 2 - linear regression
doAnalysis("linReg")

}
\seealso{
\code{\link[stats]{t.test}}, \code{\link[stats]{glm}}, \code{\link[examplePack]{doPlot}}
}
