% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doPlot}
\alias{doPlot}
\title{Does the plot for the exercise for Session 6 from the MLW / KUHeS Statistics and R workshop.}
\usage{
doPlot(analysisResult)
}
\arguments{
\item{analysisResult}{The output from the function \code{doAnalysis}.}
}
\value{
Produces a plot, but returns no other output.
}
\description{
This is an example package for Session 6 of the MLW / KUHeS Statistics and R workshop. This function plots either a box & jitter plot or a scatter plot with a linear regression model.
}
\examples{
## Example
res<-doAnalysis("t.test")
doPlot(res)

}
\seealso{
\code{\link[examplePack]{doAnalysis}}
}
